job-finder-worker/requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0
feedparser>=6.0.10

# Data Processing
pandas>=2.1.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Database (optional)
sqlalchemy>=2.0.0

# AI Integration
anthropic>=0.40.0
openai>=1.54.0
pydantic>=2.10.0
tiktoken>=0.8.0

# Firebase/Firestore Integration
firebase-admin>=6.5.0
google-cloud-logging>=3.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0
bandit[toml]>=1.7.0
pre-commit>=3.0.0
